Remove long unit test
This heuristic unit test was used to test dense schedules that could not be tested with ExactSolver. This test took about 40 seconds to run and did not add much value as the other unit tests have already verified constraints (with csp check) and verified small bqms (with Exact Solver). The only difference is that this test dealt with dense schedules. However, since the same CSP principles are applied and were tested in other unit tests (i.e. schedules), I am removing this long and somewhat redundant heurisitic unit test