Add maxgap to package requirements
This is a workaround to get CircleCI tests for py3.8 on an x86 platform for pass. Specifically, we currently have circleci with a 32-bit python (to mimic x86) on a 64-bit machine. However, since ortools is installed based on the system, a 64-bit windows or-tools is being installed and then failing on our 32-bit python 3.8. This becomes an issue if penaltymodel-mip gets used in the tests as it depends on or-tools. The workaround is to let mip fail and have penaltymodel-maxgap to fall back on.