Fix auxiliary var handling in demo.py postprocessing
With max_time=5, auxiliary variables are added to the BQM.  These were
not handled in the solution post-processing and could produce a
ValueError as noted in Issue #9.

The fix is simply to skip the auxiliary variables when extracting the
solution in demo.py.  To simplify, a utility function
is_auxiliary_variable has been added for reuse in
test_job_shop_scheduler.py.